Adding in Cool Down at end of Recommended UBL GCode.
I blindly copy and pasted this into my OctoPi Plugin, and it didn't perform as I was expecting. I made these changes because I felt it was "Safer" and "More Accurate" for me. 

Wanted to raise a PR to give back something so maybe the next person wouldn't scratch their head to why the heated bed was still on a few minutes after bed leveling.